Refactored engine management: move UUIDMismatchError and EngineManager to GRPCEngineManager
